Indices must be ints

+++ Code/get_merger_info.py
@@ -48,8 +48,8 @@ def make_pairs(filepath, reg, snap, d=0.050):
         else:
             desc_groups = np.zeros(0)
             desc_subgroups = np.zeros(0)
-            desc_pointers = np.zeros(len(masses))
-            ndescs = np.zeros(len(masses))
+            desc_pointers = np.zeros(len(masses), dtype=int)
+            ndescs = np.zeros(len(masses), dtype=int)
 
     # Filter for "resolved" galaxies
     mask = masses > 10**8
